requirements.txt
streamlit==1.24.1
Pillow==9.5.0
numpy==1.24.3
tensorflow==2.13.0
tensorflow-datasets
matplotlib==3.7.2
keras==2.13.1
pyvirtualdisplay==2.2